refactor(web-client,services): extract AlgodService from EnclaveService
