+++ web-client/src/app/services/enclave/enclave.service.ts
@@ -1,8 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import algosdk from 'algosdk';
 import { lastValueFrom } from 'rxjs';
-import { defined } from 'src/app/utils/errors/panic';
 import { environment } from 'src/environments/environment';
 import {
   CreateWallet,
@@ -12,16 +10,10 @@ import {
   SignTransaction,
   SignTransactionResult,
 } from 'src/schema/actions';
-import {
-  makePaymentTxnHelper,
-  OptionalParameters,
-  RequiredParameters,
-} from 'src/schema/algorand.helpers';
 import { AttestationReport } from 'src/schema/attestation';
 import { PublicKey, TweetNaClCrypto } from 'src/schema/crypto';
 import { from_msgpack_as } from 'src/schema/msgpack';
 import { seal_msgpack_as, unseal_msgpack_as } from 'src/schema/sealing';
-import { TransactionConfirmation, waitForConfirmation } from '../algosdk.utils';
 
 @Injectable({
   providedIn: 'root',
@@ -70,52 +62,6 @@ export class EnclaveService {
     >(walletRequest);
     const { SignTransaction: result } = response;
     return result;
-  }
-
-  async getBalance(address: string): Promise<number> {
-    const algodClient = this.getAlgodClient();
-    const accountDetails = await algodClient.accountInformation(address).do();
-    // https://developer.algorand.org/docs/reference/rest-apis/algod/v2/#account
-    return accountDetails.amount as number;
-  }
-
-  // Algorand network interface functions:
-
-  async createUnsignedTransaction(
-    required: RequiredParameters,
-    optional?: OptionalParameters
-  ) {
-    const algodClient = this.getAlgodClient();
-    const suggestedParams = await algodClient.getTransactionParams().do();
-    console.log('createUnsignedTransaction', 'got:', { suggestedParams });
-    const transaction = makePaymentTxnHelper(
-      suggestedParams,
-      required,
-      optional
-    );
-    console.log('createUnsignedTransaction', 'created:', { transaction });
-    return transaction;
-  }
-
-  async submitSignedTransaction(
-    signedTxn: Uint8Array
-  ): Promise<{ txId: string }> {
-    return await this.getAlgodClient().sendRawTransaction(signedTxn).do();
-  }
-
-  async waitForTransactionConfirmation(
-    txId: string
-  ): Promise<TransactionConfirmation> {
-    // TODO: Report rejection and timeout in a way the UI can use.
-    return waitForConfirmation(this.getAlgodClient(), txId, 4);
-  }
-
-  /** Combine {@link submitSignedTransaction} and {@link waitForTransactionConfirmation}. */
-  async submitAndConfirmTransaction(
-    signedTxn: Uint8Array
-  ): Promise<TransactionConfirmation> {
-    const { txId } = await this.submitSignedTransaction(signedTxn);
-    return await this.waitForTransactionConfirmation(txId);
   }
 
   // HTTP helpers
@@ -176,14 +122,6 @@ export class EnclaveService {
 
   protected getWalletApiUrl(path: string): string {
     return new URL(path, environment.nautilusWalletServer).toString();
-  }
-
-  protected getAlgodClient() {
-    const algod = defined(
-      environment.algod,
-      'environment.algod not configured'
-    );
-    return new algosdk.Algodv2(algod.token, algod.baseServer, algod.port);
   }
 }
